backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
pymongo==4.6.0
openai==1.3.0
python-dotenv==1.0.0
eventlet==0.33.3
python-dateutil==2.8.2

